refactor: split out journal types and type-guards
