+++ packages/core/src/new-api/internal/journal/type-guards/index.ts
@@ -1,0 +1,4 @@
+/* eslint-disable import/no-unused-modules */
+export * from "./future-level-journal-message";
+export * from "./run-level-message";
+export * from "./transaction-level-journal-messages";
