+++ packages/core/src/new-api/internal/journal/types/index.ts
@@ -1,0 +1,5 @@
+/* eslint-disable import/no-unused-modules */
+export * from "./future-level-journal-message";
+export * from "./journal";
+export * from "./run-level-journal-message";
+export * from "./transaction-level-journal-message";
